requirements.txt
os
flask=1.0.2
Flask-PyMongo=2.2.0
Werkzeug=0.14.1	
